# Java/swagger.yml
swagger: "2.0"
info:
  title: Swagger Hire-it
  version: '1.0'
...
paths:
...
  /demandas:
    post:
      tags:
      - Demanda
      summary: "Cria uma demanda"
      description: "Cria uma demanda com tecnologias e softskills"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Demanda"
      responses:
        '201':
          description: Created
        '422':
          description: Informações inválidas
    get:
      tags:
      - Demanda
      summary: "Retorna todas demandas"
      description: "Retorna todas demandas"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/DemandaResponse"
        '204':
          description: No content
  /demandas/{idDemanda}:
    get:
        tags:
        - Demanda
        summary: "Retorna uma demanda"
        description: "Retorna uma demanda de acordo com o ID no Path"
        parameters:
        - name: "idDemanda"
          in: "path"
          description: "ID da demanda que deve retornar"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/DemandaIdResponse"
          '204':
            description: No content
  /demandas/usuario/{idUsuario}:
    get:
        tags:
        - Demanda
        summary: "Retorna todas demandas de um usuário"
        description: "Retorna todas demandas de um usuário com o ID no Path"
        parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuario"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/DemandaUsuarioResponse"
          '204':
            description: No content
...
  /ofertas/usuario/{idUsuario}:
    get:
      tags:
      - Oferta
      summary: "Retorna todas ofertas de um usuário"
      description: "Retorna todas ofertas de um usuário"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/OfertaUsuarioResponse"
        '204':
          description: No content
...
  /softskills:
    get:
      tags:
      - Softskill
      summary: "Retorna todas as softskills cadastradas"
      description: "Retorna todas as softskills cadastradas"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/SoftskillsResponse"
        '204':
          description: No content
  /tecnologias:
    get:
      tags:
      - Tecnologia
      summary: "Retorna todas as tecnologias cadastradas"
      description: "Retorna todas as tecnologias cadastradas"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/TecnologiaResponse"
        '204':
          description: No content
  /tecnologias/oferta/{idOferta}:
    get:
      tags:
      - Tecnologia
      summary: "Retorna todas as tecnologias de uma oferta"
      description: "Retorna todas as tecnologias de uma oferta"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/TecnologiaOfertaResponse"
        '204':
          description: No content
  /tecnologias/oferta:
    post:
      tags:
      - Tecnologia
      summary: "Adiciona uma tecnologia de uma oferta"
      description: "Adiciona uma tecnologia de uma oferta"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Tecnologia"
      responses:
        '201':
          description: Created
        '422':
          description: Informações inválidas
  /usuarios:
    post:
      tags:
      - Usuário
      summary: "Cria um usuário"
      description: "Cria um usuário"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Usuario"
      responses:
        '201':
          description: Created
        '422':
          description: Informações inválidas
    get:
      tags:
      - Usuário
      summary: "Retorna todos usuário"
      description: "Retorna todos usuário"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/UsuarioResponse"
        '204':
          description: No content
  /usuario/{idUsuario}:
    get:
        tags:
        - Usuário
        summary: "Retorna um usuário"
        description: "Retorna um usuário de acordo com o ID no Path"
        parameters:
        - name: "idUsuário"
          in: "path"
          description: "ID do usuário que deve retornar"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/UsuarioIdResponse"
          '204':
            description: No content
    delete:
      tags:
        - Usuário
      summary: "Deleta um usuário com base em um ID"
      description: "Deleta um usuário com base em um ID"
      parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuário que deve ser deletado"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
    put:
      tags:
        - Usuário
      summary: "Atualiza um usuário com base em um ID"
      description: "Atualiza um usuário com base em um ID"
      parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuário que deve ser atualizado"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/Usuario"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
        '422':
          description: Informações inválidas
  /usuario/login:
    post:
      tags: 
      - Usuário
      summary: "Realiza o login"
      description: "Realiza o login"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UsuarioLogin"
      responses:
        '200':
          description: Login realizado com sucesso
        '400':
          description: Credenciais erradas
  /usuario/email/{email}:
    get:
      tags:
      - Usuário
      summary: "Retorna o usuário pelo email"
      description: "Retorna o usuário pelo email"
      parameters:
        - name: "email"
          in: "path"
          description: "Email que deve retornar"
          required: true
          type: "string"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/Usuario"
        '204':
          description: No content
definitions:
...
    Demanda:
      type: "object"
      properties:
        demandas:
          properties:
            titulo:
              type: "string"
              example: "Primeira demanda"
            descricao:
              type: "string"
              example: "Essa é a primeira demanda"
            salario:
              type: "number"
              example: 5000.0
              format: "double"
            usuario:
              properties:
                idUsuario:
                  type: "number"
                  example: 1
        tecnologias:
          type: array
          example: [
          {
            tempoExperiencia: 10,
            tecnologias: {
              idTecnologia: 1
            }
          },
          {
            tempoExperiencia: 10,
            tecnologias: {
              idTecnologia: 2
            }
          }
          ]
        softskills:
          type: array
          example: [
          {
            notaSoftskill: 50,
            softskill: {
              idSoftSkills: 1
            }
          },
          {
            notaSoftskill: 50,
            softskill: {
              idSoftSkills: 2
            }
          }
          ]
    DemandaResponse:
      type: "array"
      example: [
        {
          idDemanda: 1,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:18.553",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade: "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        },
        {
          idDemanda: 2,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:21.737",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade": "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        }
        ]
    DemandaIdResponse:
      type: "object"
      properties:
        idDemanda: 
          type: number
          example: 3
        titulo: 
          type: string
          example: Segunda demanda
        descricao: 
          type: string
          example: Essa é a primeira demanda
        createdAt: 
          type: string
          example: 2021-06-05T16:27:57.36
        salario: 
          type: number
          example: 2000
        usuario: 
          type: object
          properties: 
            idUsuario: 
              type: number
              example: 3
            nome: 
              type: string
              example: Thiago Barbosa
            email: 
              type: string
              example: thiaguinho@gmail.com
            senha: 
              type: string
              example: senha123
            descricao: 
              type: string
              example: DBA
            classificacao: 
              type: number
              example: 5
            telefone: 
              type: string
              example: 962345673
            localizacao: 
              type: object
              properties: 
                idLocalizacao: 
                  type: number
                  example: 1
                cidade: 
                  type: string
                  example: São Caetano do Sul
                uf: 
                  type: string
                  example: SP
                cep: 
                  type: string
                  example: 09572520
    DemandaUsuarioResponse:
      type: "array"
      example: [
        {
          idDemanda: 1,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:18.553",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade: "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        },
        {
          idDemanda: 2,
          titulo: "Primeira demanda",
          descricao: "Essa é a primeira demanda",
          createdAt: "2021-06-04T14:51:21.737",
          salario: 0.0,
          usuario: {
              idUsuario: 3,
              nome: "Thiago Barbosa",
              email: "thiaguinho@gmail.com",
              senha: "senha123",
              descricao: "DBA",
              classificacao: 5.0,
              telefone: "962345673",
              localizacao: {
                  idLocalizacao: 1,
                  cidade": "São Caetano do Sul",
                  uf: "SP",
                  cep: "09572520"
              }
          }
        }
        ]
...
    OfertaUsuarioResponse:
      type: array
      example: [
        {
            "idOferta": 2,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:49:12.613",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        },
        {
            "idOferta": 3,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:50:54.98",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        }
      ]
...
    SoftskillsResponse:
      type: "array"
      example: [
        {
            "idSoftSkills": 1,
            "softSkill": "Aberto para experiências"
        },
        {
            "idSoftSkills": 2,
            "softSkill": "Conscienciosidade"
        }
      ]
    TecnologiaResponse:
      type: "array"
      example: [
        {
            "idTecnologia": 1,
            "tecnologia": "Java"
        },
        {
            "idTecnologia": 2,
            "tecnologia": "Springboot"
        }
      ]
    TecnologiaOfertaResponse:
      type: "array"
      example: [
        {
            "idTecnologiaOferta": 1,
            "tempoExperiencia": 10,
            "tecnologias": {
                "idTecnologia": 1,
                "tecnologia": "Java"
            },
            "ofertas": {
                "idOferta": 2,
                "descricao": "Essa é a primeira demanda",
                "createdAt": "2021-06-04T18:49:12.613",
                "usuario": {
                    "idUsuario": 3,
                    "nome": "Thiago Barbosa",
                    "email": "thiaguinho@gmail.com",
                    "senha": "senha123",
                    "descricao": "DBA",
                    "classificacao": 5.0,
                    "telefone": "962345673",
                    "localizacao": {
                        "idLocalizacao": 1,
                        "cidade": "São Caetano do Sul",
                        "uf": "SP",
                        "cep": "09572520"
                    }
                }
            }
        },
        {
            "idTecnologiaOferta": 3,
            "tempoExperiencia": 10,
            "tecnologias": {
                "idTecnologia": 2,
                "tecnologia": "Springboot"
            },
            "ofertas": {
                "idOferta": 2,
                "descricao": "Essa é a primeira demanda",
                "createdAt": "2021-06-04T18:49:12.613",
                "usuario": {
                    "idUsuario": 3,
                    "nome": "Thiago Barbosa",
                    "email": "thiaguinho@gmail.com",
                    "senha": "senha123",
                    "descricao": "DBA",
                    "classificacao": 5.0,
                    "telefone": "962345673",
                    "localizacao": {
                        "idLocalizacao": 1,
                        "cidade": "São Caetano do Sul",
                        "uf": "SP",
                        "cep": "09572520"
                    }
                }
            }
        }
      ]
    Tecnologia:
      type: "object"
      properties:
        tempoExperiencia: 
          type: number
          example: 10
        tecnologias: 
          type: object
          properties: 
            idTecnologia: 
              type: number
              example: 1
        ofertas: 
          type: object
          properties: 
            idOferta: 
              type: number
              example: 4
    Usuario:
      type: "object"
      properties: 
        nome: 
          type: string
          example: Mateus Gomes
        email: 
          type: string
          example: mateusgomes@gmail.com
        senha: 
          type: string
          example: senha123
        descricao: 
          type: string
          example: Backend Developer
        telefone: 
          type: string
          example: 962345673
        localizacao: 
          type: object
          properties: 
            idLocalizacao: 
              type: number
              example: 2
    UsuarioResponse:
      type: array
      example: [
        {
        "idUsuario": 3,
        "nome": "Thiago Barbosa",
        "email": "thiaguinho@gmail.com",
        "senha": "senha123",
        "descricao": "DBA",
        "classificacao": 5.0,
        "telefone": "962345673",
        "localizacao": {
            "idLocalizacao": 1,
            "cidade": "São Caetano do Sul",
            "uf": "SP",
            "cep": "09572520"
        }
        },
        {
            "idUsuario": 6,
            "nome": "Alexandre yucra",
            "email": "alexandre@gmail.com",
            "senha": "123asd",
            "descricao": "DEV",
            "classificacao": 0.0,
            "telefone": "943666677",
            "localizacao": {
                "idLocalizacao": 1,
                "cidade": "São Caetano do Sul",
                "uf": "SP",
                "cep": "09572520"
            }
        }
      ]
    UsuarioIdResponse:
      type: object
      properties:
        idUsuario: 
          type: number
          example: 3
        nome: 
          type: string
          example: Thiago Barbosa
        email: 
          type: string
          example: thiaguinho@gmail.com
        senha: 
          type: string
          example: senha123
        descricao: 
          type: string
          example: DBA
        classificacao: 
          type: number
          example: 5
        telefone: 
          type: string
          example: 962345673
        localizacao: 
          type: object
          properties: 
            idLocalizacao: 
              type: number
              example: 1
            cidade: 
              type: string
              example: São Caetano do Sul
            uf: 
              type: string
              example: SP
            cep: 
              type: string
              example: 09572520
    UsuarioLogin:
      type: object
      properties: 
        email: 
          type: string
          example: mateus@gmail.com
        senha: 
          type: string
          example: senha123